1차 백앤드 코드 통합 진행

See merge request s10-webmobile1-sub2/S10P12E203!9

+++ gawm_web_server/gawm/src/main/java/com/cute/gawm/domain/clothe/repository/ClotheDetailRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 
 public interface ClotheDetailRepository extends MongoRepository<ClotheDetail, String> {
 
-    ClotheDetail findByClotheId(long clotheId);
+    ClotheDetail findByClotheId(int clotheId);
 }
