+++ gawm_web_server/gawm/src/main/java/com/cute/gawm/domain/clothe/repository/ClotheRepository.java
@@ -3,6 +3,6 @@ package com.cute.gawm.domain.clothe.repository;
 import com.cute.gawm.domain.clothe.entity.Clothe;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ClotheRepository extends JpaRepository<Clothe,Long> {
+public interface ClotheRepository extends JpaRepository<Clothe,Integer> {
 
 }
